% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_routine_estimation_t_mort.R
\name{estimate.offset}
\alias{estimate.offset}
\title{Estimate variance in signal}
\usage{
estimate.offset(l, var, min.t = NULL, max.t = NULL)
}
\arguments{
\item{l}{A list object with each element being measurement for a sample}

\item{var}{A character string corresponding to the name of the signal used to estimate lethal time}

\item{min.t}{The time after which signal will be used to compute offset}

\item{max.t}{The time before which signal will be used to compute offset}
}
\value{
A value that can be used as an offset in \cite{estimate.LT}
}
\description{
Estimate variance in signal
}
