% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_routine_estimation_t_mort.R
\name{extract.well}
\alias{extract.well}
\title{Extract data for a single well}
\usage{
extract.well(data)
}
\arguments{
\item{data}{A \code{data.frame} containing fluorescence measurements for the 96 wells
of a plate and several time values. See \cite{read.kinetics} for a way to obtain such
a \code{data.frame} from data files produced by a Biotech plate-reader.}
}
\value{
A \code{data.frame}
}
\description{
Extract data for a single well
}
\details{
The number of measurements might differ between wavelengths. In the Biotek reader used here, this can happen if fluorescence value exceeds the maximum value of 10^6. \code{NA} are then added to the \code{data.frame}.
}
