% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_routine_estimation_t_mort.R
\docType{package}
\name{fluoSurv}
\alias{fluoSurv}
\alias{fluoSurv-package}
\title{fluoSurv: A package for estimating insect survival data from spectrophotometry measurements}
\description{
The fluoSurv package provides functions to import fluorescence data as exported from a BioTek
microplate reader and functions to estimate insect survival from these fluorescence data.
}
\section{The function to import data}{

\link{read.kinetic}
}

\section{The function to estimate survival}{

\link{estimate.LT}
}

