% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{setup}
\alias{setup}
\title{An injection experimental setup}
\format{A data frame with 96 rows and 4 variables:
\describe{
  \item{well}{Well name.}
  \item{dilution}{Dilution factor (log-transformed) of the injected culture. 1 therefore means 10 fold dilution, while LB corresponds to negative control where insects have been injected with sterile LB culture medium.}
  \item{time_injection}{Time of the injection.}
  \item{dead}{Is the insect dead at the end of the experiment?}
}}
\description{
A dataset containing the description of an experimental setup where
96 larvae of \emph{Galleria mellonella} have been injected with a culture
of the bacterium \emph{Xenorhabdus nematophila}.
}
